+++ docs/changes.rst
@@ -25,6 +25,7 @@ Not yet released.
 * Added support for defining terminology in a glossary.
 * Moved text direction toggle to get more space for visual keyboard.
 * Added option to automatically watch projects user contributed to.
+* Added check whether translation matches glossary.
 
 Weblate 4.4.2
 -------------